Fixed some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/900



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@8681 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/extra/ustdio/locbund.c
@@ -32,7 +32,7 @@ u_locbund_new(const char *loc)
   if(result == 0)
     return 0;
 
-  len = (loc == 0 ? strlen(uloc_getDefault()) : strlen(loc));
+  len = (int32_t)(loc == 0 ? strlen(uloc_getDefault()) : strlen(loc));
   result->fLocale = (char*) uprv_malloc(len + 1);
   if(result->fLocale == 0) {
     uprv_free(result);
